Add a script for opening a shell in the buildbox container

+++ internal/build/setup-environment.sh
@@ -1,10 +1,9 @@
 #!/bin/bash
 # Usage: setup-environment.sh <DISTRO> <ARCHITECTURE>
 # Setup the build environment for a specific distribution and architecture.
-#
-# Required environment variables:
-#
-#   NGINX_VERSION
+
+# Note: this script must be idempotent because the user may call it multiple
+# times from the shell, through initpbuilder.
 
 set -e
 ROOTDIR=`dirname "$0"`
@@ -12,7 +11,6 @@ ROOTDIR=`cd "$ROOTDIR/../.." && pwd`
 source "$ROOTDIR/internal/lib/library.sh"
 
 require_args_exact 2 "$@"
-require_envvar NGINX_VERSION "$NGINX_VERSION"
 DISTRO="$1"
 ARCH="$2"
 
@@ -26,4 +24,6 @@ if [[ ! -e /cache/base-$DISTRO-$ARCH.tgz ]]; then
 	run pbuilder-dist $DISTRO $ARCH create
 	run mv ~/pbuilder/$BASE_TGZ /cache/base-$DISTRO-$ARCH.tgz
 fi
-run ln -s /cache/base-$DISTRO-$ARCH.tgz ~/pbuilder/$BASE_TGZ
+if [[ ! -e ~/pbuilder/$BASE_TGZ ]]; then
+	run ln -s /cache/base-$DISTRO-$ARCH.tgz ~/pbuilder/$BASE_TGZ
+fi
